• creates Headroom.js as ahelperfunction under utils
• refactors code to make it more reusable
