+++ portfolio-v2/src/App.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState} from "react";
 import "./App.css";
 import { Route, Switch } from "react-router-dom";
 
@@ -12,13 +12,20 @@ import Social from "./components/sections/social/Social.js";
 import CloserLook from "./components/sections/closerlook/CloserLook.js";
 import Resume from "./components/sections/resume/Resume.js";
 
+import {Headroom} from './utils/index'
+
 function App() {
   const [closeNav, setCloseNav] = useState(true);
-  const [mobileNav, setMobileNav] = useState(window.addEventListener('resize', () => {
-    var newWidth = window.innerWidth
-    setMobileNav(newWidth)
-  }))
+  const [mobileNav, setMobileNav] = useState(window.innerWidth)
 
+   window.addEventListener('resize', () => {
+  var newWidth = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
+   
+  setMobileNav(newWidth)
+})
+
+
+  window.onscroll = Headroom
 
 
 
@@ -28,9 +35,8 @@ function App() {
       <Switch>
         <Route exact path="/">
           <div>
+            {/* Keeps tabs in window width in order to switch between mobile and desktop navbar */}
           {mobileNav > 850 ? (<Nav />):(<MobileNav closeNav={closeNav} setCloseNav={setCloseNav} />)}
-             {/* <Nav />
-             <MobileNav closeNav={closeNav} setCloseNav={setCloseNav} /> */}
 
             <div onClick={() => setCloseNav(true)}>
               <Home />
