+++ portfolio-v2/src/components/nav/desktopnav/Nav.js
@@ -12,26 +12,6 @@ import {navdata} from '../../../data/navdata.js'
 export default function Nav() {
   const history = useHistory();
 
-  var prevScrollpos = window.pageYOffset;
-  window.onscroll = function() {
-    var nav = document.querySelector('.headroom')
-    var currentScrollPos = window.pageYOffset;
-    if (prevScrollpos > currentScrollPos) {
-      
-      nav.classList.add('headroom--pinned')
-      nav.classList.remove('headroom--unpinned')
-
-    } else {
-      nav.classList.add('headroom--unpinned')
-      nav.classList.remove('headroom--pinned')
-      
-    }
-    console.log(nav)
-    prevScrollpos = currentScrollPos;
-  }
-
-  // displays navbar when user scrolls up / removes navbar when user scrolls down
-
   return (
     <div className='headroom'>
       <div id="home" id="nav">
