+++ portfolio-v2/src/components/nav/mobilenav/MobileNav.css
@@ -1,24 +1,5 @@
 /* == MobileNav == */
 
-.hey {
-  z-index: 1;
-  position: absolute;
-  width: 100%;
-  top: 0;
-}
-
-.headroom--pinned {
-  position: fixed;
-  top: 0;
-  transition: top 300ms linear;
-}
-
-.headroom--unpinned {
-  position: fixed;
-  top: -150px;
-  transition: top 300ms linear;
-}
-
 @media (min-width: 851px) {
   #mobileNav {
     display: none;
