+++ portfolio-v2/src/utils/index.js
@@ -3,5 +3,6 @@ import { LiveView } from "./icons/LiveView.js";
 import { SourceCode } from "./icons/SourceCode.js";
 import { Expand } from "./icons/Expand.js";
 import { Carrot } from "./icons/Carrot.js";
+import {Headroom} from '../utils/Headroom.js'
 
-export { GoBack, LiveView, SourceCode, Expand, Carrot};
+export { GoBack, LiveView, SourceCode, Expand, Carrot, Headroom};
